联调添加购物车

+++ src/main/java/MODEL/book.java
@@ -1,5 +1,6 @@
 package MODEL;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 
 public class book {
@@ -10,8 +11,8 @@ public class book {
 	private String bauthor;
 	private String bpublish;
 	private Date bdate;
-	private double bprice;
-	private double bdiscount;
+	private BigDecimal bprice;
+	private BigDecimal bdiscount;
 	private String synopsis;
 	private long stock;
 	private String sontype;
@@ -53,16 +54,16 @@ public class book {
 	public void setBdate(Date bdate) {
 		this.bdate = bdate;
 	}
-	public double getBprice() {
+	public BigDecimal getBprice() {
 		return bprice;
 	}
-	public void setBprice(double bprice) {
+	public void setBprice(BigDecimal bprice) {
 		this.bprice = bprice;
 	}
-	public double getBdiscount() {
+	public BigDecimal getBdiscount() {
 		return bdiscount;
 	}
-	public void setBdiscount(double bdiscount) {
+	public void setBdiscount(BigDecimal bdiscount) {
 		this.bdiscount = bdiscount;
 	}
 	public String getSynopsis() {
@@ -89,8 +90,8 @@ public class book {
 				+ bpublish + ", bdate=" + bdate + ", bprice=" + bprice + ", bdiscount=" + bdiscount + ", synopsis="
 				+ synopsis + ", stock=" + stock + ", sontype=" + sontype + "]";
 	}
-	public book(String iSBN, String img, String bname, String bauthor, String bpublish, Date bdate, double bprice,
-			double bdiscount, String synopsis, long stock, String sontype) {
+	public book(String iSBN, String img, String bname, String bauthor, String bpublish, Date bdate, BigDecimal bprice,
+			BigDecimal bdiscount, String synopsis, long stock, String sontype) {
 		super();
 		ISBN = iSBN;
 		this.img = img;
